backend/requirements.txt
Django>=5.1.5
djangorestframework>=3.15.2
django-cors-headers>=4.6.0
Pillow>=10.2.0  # Para manipulação de imagens
python-dotenv>=1.0.1  # Para variáveis de ambiente
requests>=2.31.0  # Para fazer requisições HTTP 